==> 04-python-pandas/Notebooks/08-AggregationAndGrouping.ipynb <==
# Aggregate Functions and Group Statistics
* Contact: Lachlan Deer, [econgit] @ldeer, [github/twitter] @lachlandeer

So far our work with Pandas as been purely about simple transformations of each row or data; whether selecting them or creating new columns out of old ones. This is rarely enough - we typically want some degree of summarization of the data; either of an entire column; or on parts of (groups) of the column in question.

This is what we will go after here:

import pandas as pd
import numpy as np

data = pd.read_csv('out_data/state_labour_statistics.csv')

## Simple Aggregation

Like NumPy, Pandas has some of the standard data aggregations:

rng = np.random.RandomState(1234567890)
series = pd.Series(rng.rand(5))
series



series.sum()

series.mean(), series.median(), series.max()

For a data-frame; these aggregations return results for each column:

df = pd.DataFrame({'A': rng.rand(5),
                   'B': rng.rand(5)})
df

df.mean()

if you want the aggregation across a row, you can specify thats what you want:

df.mean(axis='columns')

Often times, we want to see all of the common aggregates at once. Pandas provides a function `.describe()` to do this for us:

data.describe()

Some useful aggregation functions in pandas are in the table below:

[insert table]

Whilst these aggregations can provide some notions of patterns in the data - often we want to go deeper - and compute these summaries at the level of a group (like a state of a year). Pandas has the `groupby` function that will allow us to efficiently looks at aggregate statistics on subsets of the data.

## GroupBy: The Split, Apply, Combine Paradigm
The canonical example of the groupby paradigm  is best represented in a figure:

[insert figure]

In words, GroupBy does the following:
* The *split* step breaks up a DataFrame into subsets depending on the value of a specified key
* the *apply* step computes some aggregate functions within an individual group
* The *combine* step merges the results of the computations within groups back to a single dataset

We could certainly perform all these operations by ourselves, manually - but the tedium of the task combined with a high probability of error makes us immediately want to use a built in function that abstracts away from individual steps and asks the user to rather think of the operation as a whole rather than the individual steps.

As a start - let's look at the average unemployment rate by state. As a first step we must specify the name of the column we are looking to split on:

data.groupby('state')

Think of this object as an under the hood view of a `DataFrame`. Note that it does not compute anything until an aggregation is applied to it; i.e. it uses 'lazy evaluation'

Now lets get those mena unemployment rates:

data.groupby('state').unemployment_rate.mean()

Notice if we wanted the aggregation function to be applied to all of the columns, we can skip the part where we specify the column:

data.groupby('state').mean()

and if we want multiple, but not all columns:

data.groupby('state')['unemployment_rate', 'qty_unemployed'].mean()

Very Cool!

## Challenge

Try and compute statistics where the grouping is described by multiple variables:
1. Compute the average unemployment rate by state-year
2. Compute the average unemployment rate by state-year-month
3. Compute the average unemployment rate by state-year for the Carolinas (Hint: use the .filter() function after the computation)
4. Compute the average unemployment rate by state-year and use `unstack` to view the data in a way that is different from the default output

#### Partial Solution

data.groupby(['state', 'year'])['unemployment_rate'].mean().filter(like='Carolina')

data.groupby(['state', 'year'])['unemployment_rate'].mean().unstack(level=0)

## More useful GroupBy functionality

Often we want more than one aggregate statistic, for example we may want the min, mean and max by State. We can use the `.aggregate()` function to get these:

data.groupby('state')['unemployment_rate'].aggregate([np.min, np.mean, np.max])

We can dress it up a little too if we like:

data.groupby('state')['unemployment_rate'] \
        .aggregate({'Min. Unemploy.' : np.min,
                    'Mean Unemploy.' : np.mean,
                    'Max. Unemploy.' : np.max})

In the challenge above, we used filter to select out states based on a string, but there are more uses of filter than this simple one. For example, we can filter based on the group properties:

def filter_func(x):
    return x['unemployment_rate'].min() > 5

problem_states = data.groupby('state').filter(filter_func)

problem_states.state.unique()

problem_states.groupby(['state', 'year']).unemployment_rate.mean().unstack(level=0)

### Other useful functions:

GroupBy also works alongside the following functions:
* `transform()`
* `apply()`

For lack of time we leave these for you to explore.
